feat(financial): Add HasFinancialPeriodValidation trait for transaction date validation
refactor(migrations): Remove add_service_code_to_staff_table migration

feat(migrations): Create financial periods table for accounting management

fix(migrations): Add service_code field to staff table

feat(migrations): Update staff table to include service_code

chore(version): Update version.yaml for Organization plugin with financial periods

chore(plugin): Require Organization plugin in Procurement plugin

refactor(migrations): Remove service_code migrations from procurement

fix(migrations): Update foreign key constraints in delivery order items and goods receipt note items

fix(migrations): Update foreign key constraints in purchase order items and requests

chore(version): Update version.yaml for Procurement plugin with service code support

chore(plugin): Require Organization plugin in Registrar plugin

feat(migrations): Enhance document audit trails with foreign key constraints

fix(migrations): Update document registries and types with foreign key constraints

chore(version): Update version.yaml for Registrar plugin with document tracking

chore(plugin): Require Organization plugin in Workflow plugin

chore(version): Update version.yaml for Workflow plugin with initial migrations

+++ plugins/omsb/procurement/updates/create_vendors_table.php
@@ -35,7 +35,7 @@ return new class extends Migration
             $table->softDeletes();
             
             // Foreign keys
-            $table->unsignedInteger('address_id')->nullable();
+            $table->unsignedBigInteger('address_id')->nullable();
             $table->foreign('address_id')->references('id')->on('omsb_organization_addresses')->nullOnDelete();
             
             // Indexes
